update E-ACSL oracles

+++ src/plugins/e-acsl/tests/constructs/oracle/gen_printed_data.c
@@ -423,8 +423,7 @@ int main(void)
   }
   /*@ assert \let x = real_double; \true; */ ;
   /*@ assert
-      Eva: is_nan_or_infinite:
-        \is_finite((long double)0x1.fffffffffffffffep+16383L);
+      Eva: is_nan_or_infinite: \is_finite(0x1.fffffffffffffffep+16383L);
   */
   long double real_ldouble = 0x1.fffffffffffffffep+16383L;
   __e_acsl_store_block((void *)(& real_ldouble),16UL);
